+++ CHANGELOG.md
@@ -9,6 +9,7 @@ WebSocket
 * Add rd_close_ to websocket stream state
 * stream uses flat_buffer
 * accept requires a message
+* Add wstest benchmark tool
 
 API Changes: